moved all m3u8 classes into their own file
